pyproject.toml
[project]
name = "tiangong-core"
version = "0.1.0"
description = "Tiangong agent core runtime: providers, skills, sessions, flow, and CLI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "lark-oapi>=1.4.0",
    "litellm>=1.82.1",
    "pocketflow>=0.0.3",
    "prompt-toolkit>=3.0.52",
    "qq-botpy",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
tiangong = "tiangong_core.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# 避免 flat-layout 下误把顶层目录（如 `runtime/`、`sessions/`）识别成包/命名空间包
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["tiangong_core*"]
exclude = ["runtime*", "sessions*"]

[tool.setuptools.package-data]
tiangong_core = ["templates/**/*.md"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.12.0",
    "mypy>=1.17.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
# 代码库当前包含少量超长行（多为字符串/注释/配置拼接），先不强制重排以避免无意义 diff。
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
